# src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Prestito
  description: Prestito
  version: 1.0.0
tags:
  - File

paths:
  /v1/files:
    get:
      description: return all files
      tags:
        - File
      operationId: getFiles
      responses:
        '200':
          description: a list of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileDto'

  /v1/files/{fileId}:
    get:
      description: return a file
      tags:
        - File
      operationId: getFile
      parameters:
        - $ref: '#/components/parameters/FileIdParameter'
      responses:
        '200':
          description: a file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDto'

    put:
      description: update a file
      tags:
        - File
      operationId: changeFile
      parameters:
        - $ref: '#/components/parameters/FileIdParameter'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileDto'
      responses:
        '201':
          description: changed

  /v1/files/{fileId}/thumbnail:
    get:
      description: return a thumbnail of a file
      tags:
        - File
      operationId: getThumbnail
      parameters:
        - $ref: '#/components/parameters/FileIdParameter'
      responses:
        '200':
          description: a thumbnail
          content:
            image/jpeg:
              schema:
                type: string
                format: binary

  /v1/files/{fileId}/tags:
    get:
      description: return all tags for file
      tags:
        - File
        - Tag
      operationId: getTagsForFile
      parameters:
        - $ref: '#/components/parameters/FileIdParameter'
      responses:
        '200':
          description: a list of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagDto'

    put:
      description: change tags for file
      tags:
        - File
        - Tag
      operationId: changeTagsForFile
      parameters:
        - $ref: '#/components/parameters/FileIdParameter'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/TagDto'
      responses:
        '200':
          description: changed

  /v1/tags:
    get:
      description: return all tags
      tags:
        - Tag
      operationId: getTags
      responses:
        '200':
          description: a list of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagDto'


components:
  parameters:
    FileIdParameter:
      name: fileId
      in: path
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/FileId'

    TagIdParameter:
      name: tagId
      in: path
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TagId'

  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDto'

  schemas:
    FileId:
      type: integer
      format: int64
      readOnly: true

    TagId:
      type: integer
      format: int64
      readOnly: true

    Date:
      type: string
      format: date-time
      readOnly: true

    FileDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/FileId'
        created_date:
          $ref: '#/components/schemas/Date'
        modified_date:
          $ref: '#/components/schemas/Date'
        title:
          type: string
        notes:
          type: string
        path:
          type: string
          readOnly: true
        size:
          type: integer
          readOnly: true
        created:
          $ref: '#/components/schemas/Date'

    TagDto:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/TagId'
        created_date:
          $ref: '#/components/schemas/Date'
        modified_date:
          $ref: '#/components/schemas/Date'
        title:
          type: string